backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
django-filter>=24.0
psycopg2-binary>=2.9,<3.0
Pillow>=10.4,<11.0
celery>=5.4,<6.0
redis>=5.0,<6.0
django-celery-beat>=2.6,<3.0
gunicorn>=22.0,<23.0
python-dotenv>=1.0,<2.0
